Make usernames truely unqiue using a player repository
Added a generic BaseRepository class to follow the M.A.D pattern. Had to create
a simple extension of LiveData called FirebaseLiveData which removes itself as a
listener when inactivity to prevent memory leaks. Created a player repo to check
generated usernames against.

trying to follow the general M.A.D pattern so repo's should act as a single source
of truth.