Add live updating user scores to the home screen #8
An inner score class that holds high,low,and total scores. This
reduces the number of loops require when handling a change from the post
repository.

The new getWhereEqual method which returns a list of the entities with a matching
field and value. This improves subclass ergonomics and readability. getPostByPlayer
is the first method make use of getWhereEqual

closes: #8