Bug: fix stuck on prev Player bug
Most of the fetches for data happened in OnCreateView. OnCreateView is
not called when a saved fragment is loaded, this meant when signing out
the prevous user was still being listened to. Starting a new main
activity intent in the welcome screen was choosen over moving fetches
to OnStart because it improves performance