Add post screen (#90)
* Add post screen

Safe args creates a kotlin version conflict. Instead target was updated
to 33, to use bundle and a seriziable entity.

* working dirty commit

* Add scores collection

A new collection allows the post repository to keep track of a players
score, and which qrcodes they have scanned, without interfering with
the player repo. Need for see who else scanned.